testing network pass

+++ app/src/main/java/dev/android/kevin/project/model/DetailBean.java
@@ -48,7 +48,7 @@ public class DetailBean {
 
         private int price_level;
 
-        private int rating;
+        private float rating;
 
         private List<Reviews> reviews;
 
@@ -124,11 +124,11 @@ public class DetailBean {
             this.price_level = price_level;
         }
 
-        public int getRating() {
+        public float getRating() {
             return rating;
         }
 
-        public void setRating(int rating) {
+        public void setRating(float rating) {
             this.rating = rating;
         }
